feat(rust): use private address for hardy
