+++ rust-bindings/hardy-cspcl/Cargo.toml
@@ -22,3 +22,4 @@ tokio = { version = "1.51.1", features = ["time"] }
 serde = { version = "1.0.228", features = ["derive"], optional = true }
 cspcl-bindings = { path = "../cspcl", package = "cspcl" }
 futures-util = "0.3.32"
+bytes = "1.12.0"
